# api/openapi.yaml
openapi: 3.1.0
info:
  title: portal API
  version: 1.0.0
  description: |
    The portal HTTP API. Every route chi serves under /api/v1
    (internal/server/server.go) is described here; the schemas mirror the
    handler-layer *Response types in internal/handler (the wire boundary —
    repository rows never serialize directly).

    web/src/api/types.ts is generated from this file with openapi-typescript
    (`npm run generate:api` in web/); openapi-fetch consumes the generated
    `paths` interface. Edit this spec, regenerate, and the frontend contract
    follows — CI fails on drift.

    Two surfaces sit outside the JSON request/response model:
    - GET /workspaces/{workspaceId}/runs/{runId}/logs/ws upgrades to a
      WebSocket (documented on the operation; the stream itself is
      newline-delimited log text, authenticated via the
      `bearer` + token subprotocol pair).
    - GET /auth/github and /auth/github/callback are browser redirect hops of
      the OAuth flow, not XHR endpoints.

    Errors use one envelope everywhere (ErrorResponse) via respond.Error /
    respond.FromError; operations declare it as their default response.
servers:
  - url: /api/v1
paths:
...
  /workspaces/{workspaceId}/runs:
    get:
      operationId: listRuns
      summary: List a workspace's runs
      parameters:
        - { name: workspaceId, in: path, required: true, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: per_page, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Runs
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunList" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createRun
      summary: Create a run (role elevates per operation; apply/destroy on a workspace that requires approval needs admin)
      parameters:
        - { name: workspaceId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRunRequest" }
      responses:
        "201":
          description: Created run
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        default: { $ref: "#/components/responses/Error" }
...
  /workspaces/{workspaceId}/runs/{runId}/cancel:
    post:
      operationId: cancelRun
      summary: Cancel a run (operator)
      parameters:
        - { name: workspaceId, in: path, required: true, schema: { type: string } }
        - { name: runId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Cancelled run
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: The standard error envelope (respond.Error / respond.FromError).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string, description: "HTTP status text." }
        message: { type: string }
        request_id: { type: string, description: "Set on 5xx responses so logs can be correlated." }
      required: [error]
...
    RunStatus:
      type: string
      enum:
        - pending
        - queued
        - planning
        - planned
        - awaiting_approval
        - applying
        - applied
        - errored
        - cancelled
        - discarded
    RunOperation:
      type: string
      enum: [plan, apply, destroy, import, test]
    Run:
      type: object
      properties:
        id: { type: string }
        workspace_id: { type: string }
        org_id: { type: string }
        operation: { $ref: "#/components/schemas/RunOperation" }
        status: { $ref: "#/components/schemas/RunStatus" }
        plan_output: { type: string }
        plan_log_url: { type: string }
        apply_log_url: { type: string }
        resources_added: { type: integer }
        resources_changed: { type: integer }
        resources_deleted: { type: integer }
        error_message: { type: string }
        commit_sha: { type: string }
        plan_json_url: { type: string }
        created_by: { type: string }
        started_at: { type: [string, "null"], format: date-time }
        finished_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, workspace_id, org_id, operation, status, created_by, created_at, updated_at]
    CreateRunRequest:
      type: object
      properties:
        operation: { $ref: "#/components/schemas/RunOperation" }
        imports:
          type: array
          items:
            type: object
            properties:
              address: { type: string }
              id: { type: string }
            required: [address, id]
...
    RunList:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Run" }
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
      required: [data, total, page, per_page]
...
security:
  - bearerAuth: []